docker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
transformers==4.46.3
accelerate==1.1.1
pillow==11.0.0
timm==1.0.12
einops==0.8.0
sentencepiece==0.2.0
python-multipart==0.0.17
